server_python/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
openai-agents>=0.9.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
python-dotenv==1.0.1
httpx==0.28.1
feedparser==6.0.11
vaderSentiment==3.3.2
google-auth==2.37.0
google-api-python-client==2.157.0
google-auth-httplib2==0.2.0
google-api-core==2.19.2
python-multipart==0.0.20
